+++ src/main/java/com/reason/ide/search/reference/ORPsiLowerSymbolReference.java
@@ -48,7 +48,6 @@ public class ORPsiLowerSymbolReference extends ORMultiSymbolReference<RPsiLowerS
         }
 
         // Gather instructions from element up to the file root
-        //Deque<PsiElement> instructionsOLD = ORReferenceAnalyzer.createInstructionsOLD(myElement, true, myTypes);
         Deque<PsiElement> instructions = ORReferenceAnalyzer.createInstructionsBackward(myElement, true, myTypes);
 
         // Test if source element is part of a record/object chain